check last buyer message

+++ routes/console.php
@@ -13,18 +13,22 @@ Schedule::call(function () {
 
     if (isset($optNotification->is_active) && $optNotification->is_active == true) {
         $fbChecker = new Feedback();
+        $notificationService = new ServicesNotification();
 
-        Notification::where('is_notificated', false)->chunk(300, function (Collection $notificationList) use ($optNotification, $fbChecker) {
+        Notification::where('is_notificated', false)->chunk(300, function (Collection $notificationList) use ($optNotification, $fbChecker, $notificationService) {
             foreach ($notificationList as $notification) {
                 // Проверим: есть отзыв в заказе или нет. Если false, то отсылаем сообщение, иначе пропускаем.
                 if ($fbChecker->check($notification->invoice_id) === false) {
                     $timeNotification = $notification->time_of_purchase + $optNotification->time_of_sending;
 
-                    if ($timeNotification < time()) {
-                        $notificationService = new ServicesNotification();
-                        $notificationService->sendMessage(false, $notification->invoice_id);
+                    // Проверим: есть ли сообщение от клиента, которое не прочитано нами
+                    if ($notificationService->checkSeenMessage($notification->invoice_id)) {
+                        if ($timeNotification < time()) {
 
-                        $notification->delete();
+                            $notificationService->sendMessage(false, $notification->invoice_id);
+
+                            $notification->delete();
+                        }
                     }
                 }
             }
